moved some functions and updated tests
